+++ src/main/java/oncoding/concoder/etc/Language.java
@@ -1,0 +1,12 @@
+package oncoding.concoder.etc;
+
+public enum Language {
+    PYTHON("python"),
+    CPP("cpp");
+
+    private final String lang;
+
+    Language(String lang) {
+        this.lang = lang;
+    }
+}
